chore: fix temperature type/usage, update CLAUDE.md, remove orphan vault/
- gatekeeper.py: temperature stored as float (was int), matches LLM API spec
- base_agent.py: pass temperature from config to every messages.create() call
  (was stored in config but never forwarded — dead config value)
- CLAUDE.md: update test count 104→130, add ApiGatekeeper + screenshots to
  grading checklist, add 3 new test files to test table
- Remove vault/ (untracked orphan directory with only .obsidian/ settings)

+++ src/agents/gatekeeper.py
@@ -30,7 +30,7 @@ class AgentConfig:
         agents = cfg["agents"]
         self.model: str = agents["model"]
         self.max_tokens_per_call: int = agents["max_tokens_per_call"]
-        self.temperature: int = agents.get("temperature", 0)
+        self.temperature: float = float(agents.get("temperature", 0))
         self.max_tokens: dict[str, int] = agents.get("max_tokens", {})
         self.iteration_sleep_seconds: int = agents.get("iteration_sleep_seconds", 1)
